+++ android/app/src/androidTest/java/com/oppzippy/openscq30/hilt/MockSoundcoreDeviceFactoryModule.kt
@@ -1,7 +1,7 @@
 package com.oppzippy.openscq30.hilt
 
-import com.oppzippy.openscq30.features.soundcoredevice.SoundcoreDeviceFactoryModule
-import com.oppzippy.openscq30.features.soundcoredevice.api.SoundcoreDeviceFactory
+import com.oppzippy.openscq30.features.soundcoredevice.api.SoundcoreDeviceConnector
+import com.oppzippy.openscq30.features.soundcoredevice.dagger.SoundcoreDeviceConnectorModule
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.components.SingletonComponent
@@ -12,12 +12,12 @@ import javax.inject.Singleton
 @Module
 @TestInstallIn(
     components = [SingletonComponent::class],
-    replaces = [SoundcoreDeviceFactoryModule::class],
+    replaces = [SoundcoreDeviceConnectorModule::class],
 )
 object MockSoundcoreDeviceFactoryModule {
     @Provides
     @Singleton
-    fun provideSoundcoreDeviceFactoryProvider(): SoundcoreDeviceFactory {
+    fun provideSoundcoreDeviceFactoryProvider(): SoundcoreDeviceConnector {
         return mockk()
     }
 }