+++ android/app/src/main/java/com/oppzippy/openscq30/features/soundcoredevice/api/SoundcoreDeviceConnector.kt
@@ -2,8 +2,8 @@ package com.oppzippy.openscq30.features.soundcoredevice.api
 
 import kotlinx.coroutines.CoroutineScope
 
-interface SoundcoreDeviceFactory {
-    suspend fun createSoundcoreDevice(
+interface SoundcoreDeviceConnector {
+    suspend fun connectToSoundcoreDevice(
         macAddress: String,
         scope: CoroutineScope,
     ): SoundcoreDevice?