+++ android/app/src/main/java/com/oppzippy/openscq30/features/soundcoredevice/impl/BluetoothDeviceFinder.kt
@@ -1,0 +1,9 @@
+package com.oppzippy.openscq30.features.soundcoredevice.impl
+
+import android.bluetooth.BluetoothDevice
+
+interface BluetoothDeviceFinder {
+    fun findByMacAddress(
+        macAddress: String,
+    ): BluetoothDevice?
+}