+++ android/app/src/main/java/com/oppzippy/openscq30/features/soundcoredevice/impl/SoundcoreDeviceCallbackHandlerFactory.kt
@@ -1,0 +1,7 @@
+package com.oppzippy.openscq30.features.soundcoredevice.impl
+
+import android.content.Context
+
+interface SoundcoreDeviceCallbackHandlerFactory {
+    fun createSoundcoreDeviceCallbackHandler(context: Context): SoundcoreDeviceCallbackHandler
+}