+++ android/app/src/main/java/com/oppzippy/openscq30/features/soundcoredevice/impl/SoundcoreDeviceCallbackHandlerFactoryImpl.kt
@@ -1,0 +1,11 @@
+package com.oppzippy.openscq30.features.soundcoredevice.impl
+
+import android.content.Context
+import javax.inject.Inject
+
+class SoundcoreDeviceCallbackHandlerFactoryImpl @Inject constructor() :
+    SoundcoreDeviceCallbackHandlerFactory {
+    override fun createSoundcoreDeviceCallbackHandler(context: Context): SoundcoreDeviceCallbackHandler {
+        return SoundcoreDeviceCallbackHandler(context)
+    }
+}